requirement.txt
sounddevice
openai
requests
scipy
google-cloud-storage
shutil
opencv-python
backgroundremover